Refactor [form schema] group form
